Midterm/requirements.txt
fastapi
uvicorn
langchain
openai
qdrant-client
python-dotenv